Register download notification when client process is exited
[issue#] N_SE-46078
[cause] It didn't to check download notification when client process is exited
[solution] At this case, the download process register "Failed" notification to Quick Panel.


	modified:   provider/download-provider-notification.c
	modified:   provider/download-provider-thread-request.c

+++ provider/download-provider-notification.c
@@ -567,7 +567,7 @@ void dp_update_downloadinginfo_notification(int priv_id, double received_size, d
 
 	if (file_size > 0) {
 		double progress;
-		progress = received_size / file_size;
+		progress = (double)received_size / (double)file_size;
 		err = notification_update_progress(NULL, priv_id, progress);
 		if (err != NOTIFICATION_ERROR_NONE)
 			TRACE_ERROR("[FAIL] update noti progress[%s]",
